Promote ID generation to storage package
