+++ storage/storage.go
@@ -2,6 +2,10 @@ package storage
 
 import (
 	"context"
+	"crypto/rand"
+	"encoding/base32"
+	"io"
+	"strings"
 	"time"
 
 	"github.com/golang/protobuf/proto"
@@ -51,4 +55,19 @@ type errConflict interface {
 func IsConflictErr(err error) bool {
 	_, ok := err.(errConflict)
 	return ok
+}
+
+// Kubernetes only allows lower case letters for names.
+//
+// TODO(ericchiang): refactor ID creation onto the storage.
+var encoding = base32.NewEncoding("abcdefghijklmnopqrstuvwxyz234567")
+
+// NewID returns a random string which can be used as an ID for objects.
+func NewID() string {
+	buff := make([]byte, 16) // 128 bit random ID.
+	if _, err := io.ReadFull(rand.Reader, buff); err != nil {
+		panic(err)
+	}
+	// Avoid the identifier to begin with number and trim padding
+	return string(buff[0]%26+'a') + strings.TrimRight(encoding.EncodeToString(buff[1:]), "=")
 }
